+++ config/connection.js
@@ -6,7 +6,7 @@ const sequelize = new Sequelize(
   process.env.DB_USER,
   process.env.DB_PASSWORD,
   {
-    host: "localhost",
+    host: "127.0.0.1",
     dialect: "mysql",
     port: 3306,
   }
